+++ Scripts/Enemies/IEnemy.cs
@@ -2,6 +2,6 @@ namespace ShiftingSepulcher
 {
     public interface IEnemy
     {
-        bool IsDead {get;}
+        bool IsAlive {get;}
     }
 }